Fix documentation formatting
Godoc treats all indented code as code blocks, and was incorrectly
interpreting list items in doc.go as code blocks. In addition, Godoc
does not recognize the Markdown URL format, so just put links in
parentheses and omit the braces.

Finally, the code samples had an extra unnecessary level of
indentation on account of using 4 spaces for indentation instead of
tabs. Use tabs instead so they appear correctly on the godoc website.